Implement solution for deduplication of random results #7
Added Cache-Control and Etag header values for responses. Added no-store
to random sort order requests, as well as a cache bypass mechanism. This
mechanism currently only triggers when sort order is set to random.
Added calls to cache service to maintain expiration times as well as
accessor for max age.

Change-Id: If7b4b57c7e265fe69ef4fdefec4249ea55bcdf5d
Signed-off-by: Martin Lowe <martin.lowe@eclipse-foundation.org>