Create performance analytics tests #5
Added 5 tests that can be run using jmeter to test an active client.
They allow for throughput testing using 50 threads, and load tests for 1
to 4 times production average request load. They will output csv files
that can be observed for statistics on request times. These tests are
time unbounded and need to be actively stopped to end a test.

Signed-off-by: Martin Lowe <martin.lowe@eclipse-foundation.org>